Move eth endpoints into header oracle
