+++ trin-core/src/jsonrpc/mod.rs
@@ -1,5 +1,4 @@
 pub mod endpoints;
-pub mod eth;
 pub mod handlers;
 pub mod service;
 pub mod types;
